x86_64,entry: Use lretq to return to userspace when possible
IRET serializes, but LRETQ does not.  Try to use LRETQ to return
to userspace when possible.

It's possible if the saved RF and TF are clear and IF is set.

This cuts about 23ns off of the IRET-to-userspace path on my
machine.

NB: Currently broken.  The bad_lretq path will enable interrupts
too early, causing a possible crash.

Signed-off-by: Andy Lutomirski <luto@amacapital.net>