x86/entry/64, entry: Set up a valid sysenter stack and prepare for 32-bit merge
Oddly, 64-bit kernels already allocate a percpu sysenter stack, but
they don't enable it.  Enable the stack and tweak the rest of the
sysenter setup code to be similar to the 32-bit version.

This eliminates the only place in the kernel in which TF could be set
without a valid stack.

Signed-off-by: Andy Lutomirski <luto@kernel.org>