fork, arches: Allocate thread "stack", not "info"
For historical reasons, the thread stack allocation functions are
named like alloc_thread_info_node().  This is confusing: they
allocate the *stack*, and the fact that thread_info lives on the
stack (and the fact that thread_info exists at all) is mostly a
historical accident.

Rename the functions and stop using 'struct thread_info *' to refer
to the bottom of the stack in core stack allocation code.

These patches exist to help prepare the core code to make
thread_info optional.

Signed-off-by: Andy Lutomirski <luto@kernel.org>