scatterlist: Introduce sg_init_stackbuf to point at the stack
There are several users of the crypto API that create scatterlist
pointing to the stack.  This is bad for any number of reasons.

To make it compatible with vmalloced stacks, add a new scatterlist
API for this use case.  This will have the added benefit of
annotating code paths that do this.

Signed-off-by: Andy Lutomirski <luto@kernel.org>