dma-api: Teach the "DMA-from-stack" check about vmapped stacks
If we're using CONFIG_VMAP_STACK and we manage to point an sg entry
at the stack, then either the sg page will be in highmem or sg_virt
will return the direct-map alias.  In neither case will the existing
check_for_stack() implementation realize that it's a stack page.

Fix it by explicitly checking for stack pages.

Signed-off-by: Andy Lutomirski <luto@kernel.org>