nvme: Add a quirk to disable APST on a buggy Samsung device
This can't use the existing NVMe-over-PCI quirk mechanism because,
as far as I can tell, all the Samsung devices use the same PCI
device id.  Instead, this adds a "core" quirk table that matches on
the vendor id and one or more of (model name, firmware revision),
and quirks found in that table will be added to the quirk mask.

Signed-off-by: Andy Lutomirski <luto@kernel.org>