x86/fpu: copy_uabi_to_xstate shouldn't copy mxcsr_mask
mxcsr_mask is a propoerty of the CPU.  XSAVE and relatives write it to memory,
but XRSTOR and relatives don't read it.  copy_uabi_to_xstate is intended to
logically work like XRSTOR, so stop using the user-supplied mxcsr_mask.

If XFEATURE_MASK_FP is set in the incoming buffer, mxcsr_mask will still
be replaced by the user-supplied value.  The next patch will fix that.

Signed-off-by: Andy Lutomirski <luto@kernel.org>