x86/insn: Check DPL when resolving segments
The instruction decoder sometimes needs to look up a segment.  Real x86
code does not (except for actual segment loads) because it uses the
cached segments, but those cached values are hidden even to the kernel
and certainly can't be found in pt_regs.  So the decoder needs to dig
the values out from the GDT or LDT.

The decoder didn't check DPL, though.  When a CPU actually *executes* an
instruction that uses a segment, it's supposed to be impossible for the
segment to reference a DPL==0 segment when the access isn't privileged
to use that segment [0], but there is nothing that fundamentally
requires that the kernel only decodes an instruction in a context when
the CPU would actually manage to execute it.  In particular, if
something modifies a selector in pt_regs between the last time the task
successfully executed user code and when decoding an instruction, then
the CPU's segment load / revalidate logic doesn't get a chance to do any
privilege checks, and the decoder might produce bogus output if the
selector in pt_regs points to a DPL==0 segment.

So check DPL if RPL != 0 or CPL != 0.

This is all a bit odd -- the entire design of the kernel's instruction
decoder assumes that the instruction is being decoded in the context of
current -- the API doesn't even allow users to specify a task.  So
anything that decodes instructions, especially user instructions,
belonging to a different task will get bogus results.

[0] Other than conforming code segments, but Linux doesn't use those.

Signed-off-by: Andy Lutomirski <luto@kernel.org>