x86/efi: Remove the real mode trampoline in boot services quirk
The only machine known to have required the quirk that allocated the
real mode trampoline in boot services memory is my laptop, and my
laptop is also able to boot with the real mode trampline below 64kB.
The quirk bitrotted (the code involved no longer runs
unconditionally), so just remove it.

Good riddance.

Signed-off-by: Andy Lutomirski <luto@kernel.org>